feat: initial commit that ties together the xp smart contract framework this is still a wip
In this commit we expose and introduce the concept of XP,LXP, and
Decaying Soulbound Token Staking, a novel reward/reputation
framework, that provides additional incentives for staying an
active participant in a community while avoiding lock in .

The architecture supported is a translation of Lukso standards
into the diamond erc-2535 standard format. This will allow for
all contracts to be upgraded, which will allow us to expand
pricing functionality in the future.